+++ index.js
@@ -1,77 +1,91 @@
 const path = require('path');
-const EventEmitter = require('events').EventEmitter;
 const fs = require('fs');
 const {promisify} = require('util');
-const merge = require('lodash/merge');
+const readDir = promisify(fs.readdir);
+const merge = require('lodash.merge');
 const trid = require('trid');
-const forEach = require('lodash/forEach');
 const Module = require('./lib/Module');
 const logger = require('./lib/logger');
-const pack = require('./package');
+const {getParentPackageInfo} = require('./lib/util');
+const kojoPackage = require('./package');
 
 
-const readDir = promisify(fs.readdir);
+module.exports = class {
 
+    constructor(options) {
 
-module.exports = class extends EventEmitter {
-
-    constructor (name, options, packageInfo={}) {
-        super();
         const defaults = {
             subsDir: 'subscribers',
             modulesDir: 'modules',
+            parentPackage: getParentPackageInfo(),
+            name: '工場',
             icon: '☢'
         };
-        const kojo = this;
+        this._options = options;
+        this.config = this._options ? merge(defaults, this._options) : defaults;
+        const {name} = this.config;
         const id = new trid({prefix: name});
-        kojo.id = id.base();
-        kojo.name = name;
-        kojo.config = options ? merge(defaults, options) : defaults;
-        kojo._extras = {};
-        kojo._modules = {};
-        kojo._pack = packageInfo;
-        kojo._subscribers = [];
+        this.id = id.base();
+        this.name = name;
+        this._extras = {};
+        this.modules = {};
+        this._subscribers = [];
     }
 
     async ready() {
 
         const kojo = this;
-        const icon = kojo.config.icon;
+        const {icon, parentPackage} = kojo.config;
 
         console.log('*************************************************************');
-        console.log(`  ${icon} ${kojo.id}  |  ${kojo._pack.name}@${kojo._pack.version}  |  ${pack.name}@${pack.version}`);
+        console.log(`  ${icon} ${kojo.id}  |  ${parentPackage.name}@${parentPackage.version}  |  ${kojoPackage.name}@${kojoPackage.version}`);
         console.log('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~');
-        process.stdout.write(`    ${icon} loading modules...`);
+
+        // MODULES
+
         const modulesDir = path.join(process.cwd(), kojo.config.modulesDir);
-        const moduleDirs = await readDir(modulesDir);
         try {
-            forEach(moduleDirs, (moduleDir) => {
+            const moduleDirs = await readDir(modulesDir);
+            process.stdout.write(`    ${icon} loading modules...`);
+            moduleDirs.forEach((moduleDir) => {
                 const modulePath = path.join(modulesDir, moduleDir);
                 if (!fs.lstatSync(modulePath).isDirectory()) return;
                 const moduleName = path.basename(modulePath);
-                kojo._modules[moduleName] = new Module(moduleName, modulePath, kojo);
+                kojo.modules[moduleName] = new Module(moduleName, modulePath, kojo);
             });
             console.log('done');
         } catch (error) {
-            console.log('error');
-            throw error;
+            if (error.code === 'ENOENT' && error.path === modulesDir && !kojo._options.modulesDir)
+                console.log(`    ${icon} skipping modules`);
+            else {
+                console.log('error');
+                throw error;
+            }
         }
 
+        // SUBSCRIBERS
 
-        console.log(`    ${icon} loading subscribers`);
         const subsDir = path.join(process.cwd(), kojo.config.subsDir);
-        const subscriberFiles = await readDir(subsDir);
-        const subsDone = [];
-        forEach(subscriberFiles, async (subscriberFile) => {
-            const subName = path.basename(subscriberFile, '.js');
-            const requirePath = path.join(subsDir, subscriberFile);
-            kojo._subscribers.push(subName);
-            let subsWrapper = require(requirePath);
-            subsDone.push(subsWrapper(kojo, logger(kojo, 'sub', subName)));
-        });
-        await Promise.all(subsDone);
-        console.log(`    ${icon} ${kojo.name} ready`);
+        try {
+            const subsDone = [];
+            const subscriberFiles = await readDir(subsDir);
+            console.log(`    ${icon} loading subscribers`);
+            subscriberFiles.forEach(async (subscriberFile) => {
+                const subName = path.basename(subscriberFile, '.js');
+                const requirePath = path.join(subsDir, subscriberFile);
+                kojo._subscribers.push(subName);
+                let subsWrapper = require(requirePath);
+                subsDone.push(subsWrapper(kojo, logger(kojo, 'sub', subName)));
+            });
+            await Promise.all(subsDone);
+        } catch (error) {
+            if (error.code === 'ENOENT' && error.path === subsDir) {
+                console.log(`    ${icon} skipping subscribers`);
+            } else
+                throw error;
+        }
 
+        console.log(`    ${icon} kojo "${kojo.name}" ready`);
         console.log('*************************************************************');
     }
 
@@ -81,11 +95,6 @@ module.exports = class extends EventEmitter {
 
     get(key) {
         return key ? this._extras[key]: this._extras;
-    }
-
-    module(name) {
-        if (!this._modules[name]) throw new Error(`module '${name}' is either not loaded yet or unknown`);
-        return this._modules[name];
     }
 
 };