+++ test/test_kojo/modules/bravo/methodA.js
@@ -1,7 +1,6 @@
 module.exports = async function () {
 
     const {logger} = this;
-
     logger.debug(`called`);
     return 'bravo';
 };